move `knex/` folder to root (align with askmebot setup)
